Added dashboard styling
