+++ client/index.html
@@ -4,8 +4,10 @@
     <meta charset="UTF-8" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
-      href="https://fonts.googleapis.com/css2?family=Libre+Franklin:wght@100&display=swap"
+      href="https://fonts.googleapis.com/css2?family=Libre+Franklin:wght@100;600&display=swap"
       rel="stylesheet"
     />
     <link
