+++ client/src/components/Map/AddLocationButton.jsx
