+++ client/src/components/Map/ShowAllLocationsButton.jsx
