+++ client/src/pages/RoadtripMap.jsx
@@ -1,4 +1,4 @@
-import Map from "../components/Map";
+import Map from "../components/Map/Map";
 
 export default function Roadtrip() {
   return <Map />;
